Move CustomListItem.tsx to PanelTabs directory.
