+++ new-log-viewer/src/components/SidebarContainer/PanelTabs/CustomListItem.tsx
