+++ new-log-viewer/src/components/SidebarContainer/PanelTabs/FileInfoTab.tsx
@@ -14,7 +14,7 @@ import {
     TAB_NAME,
 } from "../../../typings/tab";
 import {formatSizeInBytes} from "../../../utils/units";
-import CustomListItem from "../../CustomListItem";
+import CustomListItem from "./CustomListItem";
 import CustomTabPanel from "./CustomTabPanel";
